1.0.0

+++ src/StackMachine/Variable.java
@@ -2,18 +2,17 @@ package StackMachine;
 
 class Variable {
 
-    private int value;
+    private float value;
     private String type;
 
     Variable(int value, String type){
         this.value = value;
         this.type = type;
     }
-    void setValue(int value){
+    void setValue(float value){
         this.value = value;
     }
-
-    int getValue(){
+    float getValue(){
         return value;
     }
     String getType(){
